requirements.txt
autogen_agentchat==0.5.6
autogen_core==0.5.6
autogen_ext==0.5.6
dlib==19.24.8
numpy==2.2.5
opencv_contrib_python==4.11.0.86
opencv_python==4.7.0.68
opencv_python==4.11.0.86
Pillow==11.2.1
pylatex==1.4.2
Requests==2.32.3
